functions/scale-down-ondemand-np/requirements.txt
google-cloud-container
google-cloud-compute